observability-be/pyproject.toml
[project]
name = "observability-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clickhouse-connect>=0.9.2",
    "fastapi>=0.121.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
